+++ pywbem_mock/config.py
@@ -39,30 +39,24 @@ __all__ = ['OBJECTMANAGERNAME', 'SYSTEMNAME',
 #: Name for the object manager It is used in defining user-defined provider
 #: properties for CIM_Namespace provider and CIM_ObjectManager provider if
 #: if they are defined.
-
 OBJECTMANAGERNAME = 'FakeObjectManager'
 
 #: Value for the CIM_ObjectManagerCreationClassName defined in the
 #: CIM_ObjectManager class.
-
 OBJECTMANAGERCREATIONCLASSNAME = 'CIM_ObjectManager'
 
 #: The name of the mock object. This string value becomes part of the
 #: CIM_ObjectNamager instance and CIM_Namespace instances
-
 SYSTEMNAME = 'MockSystem_WBEMServerTest'
 
 #: Name for the property SystemCreationClassname defined a number of
 #: CIM classes that are used by the pywbem_mock including CIM_Namespace
-
 SYSTEMCREATIONCLASSNAME = 'CIM_ComputerSystem'
 
 #: Default positive integer value for the Open... request responder default
 #: value for MaxObjectCount if the value was not provided by the request.
-
 DEFAULT_MAX_OBJECT_COUNT = 100
 
 #: Maximum timeout in seconds for pull operations if the OperationTimeout
 #: parameter is not included in the request
-
 OPEN_MAX_TIMEOUT = 40